refactor: 컨벤션 위반 패키지 구조 수정
