+++ backend/src/main/java/site/coduo/todo/controller/TodoController.java
@@ -17,12 +17,12 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import site.coduo.todo.controller.docs.TodoDocs;
-import site.coduo.todo.controller.request.CreateTodoRequest;
-import site.coduo.todo.controller.request.UpdateTodoContentRequest;
-import site.coduo.todo.controller.request.UpdateTodoOrderRequest;
-import site.coduo.todo.controller.response.GetTodoResponse;
 import site.coduo.todo.domain.Todo;
 import site.coduo.todo.service.TodoService;
+import site.coduo.todo.service.dto.CreateTodoRequest;
+import site.coduo.todo.service.dto.TodoReadResponse;
+import site.coduo.todo.service.dto.UpdateTodoContentRequest;
+import site.coduo.todo.service.dto.UpdateTodoOrderRequest;
 
 @RequiredArgsConstructor
 @RestController
@@ -31,10 +31,10 @@ public class TodoController implements TodoDocs {
     private final TodoService todoService;
 
     @GetMapping("/{accessCode}/todos")
-    public List<GetTodoResponse> getTodos(@PathVariable("accessCode") final String accessCode) {
+    public List<TodoReadResponse> getTodos(@PathVariable("accessCode") final String accessCode) {
         final List<Todo> allTodos = todoService.getAllOrderBySort(accessCode);
         return IntStream.range(0, allTodos.size())
-                .mapToObj(index -> GetTodoResponse.from(allTodos.get(index), index))
+                .mapToObj(index -> TodoReadResponse.from(allTodos.get(index), index))
                 .toList();
     }
