+++ backend/src/main/java/site/coduo/todo/service/dto/CreateTodoRequest.java
@@ -1,4 +1,4 @@
-package site.coduo.todo.controller.request;
+package site.coduo.todo.service.dto;
 
 import jakarta.validation.constraints.NotBlank;
