+++ backend/src/main/java/site/coduo/todo/service/dto/UpdateTodoContentRequest.java
@@ -1,4 +1,4 @@
-package site.coduo.todo.controller.request;
+package site.coduo.todo.service.dto;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
