+++ backend/src/main/java/site/coduo/todo/service/dto/UpdateTodoOrderRequest.java
@@ -1,4 +1,4 @@
-package site.coduo.todo.controller.request;
+package site.coduo.todo.service.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
